backend/requirements.txt
django==4.1.13
djangorestframework==3.14.0
pymongo==4.5.0
boto3==1.28.48
django-storages==1.14
python-dotenv==1.0.0
django_cors_headers==4.2.0
adrf==0.1.2
# async-poe-client==0.3.6
requests==2.29.0
djangorestframework-simplejwt==5.3.0
google-api-python-client
google-apps-meet
google-auth-httplib2
google-auth-oauthlib